Move the file creating and downloading to a separate module
